fix a lot of warnings (#50)

+++ Stratega/src/Configuration/GameConfig.cpp
@@ -155,7 +155,7 @@ namespace SGA
 		return std::move(state);
 	}
 
-	int GameConfig::getNumberOfPlayers() const
+	size_t GameConfig::getNumberOfPlayers() const
 	{
 		return numPlayers == -1 ? agentParams.size() : numPlayers;
 	}
